requirements.txt
discord.py==0.16.12
PyYAML==3.13
pytz==2018.9
